Introduce Grafana Agent Flow system packages (#3369)
* make: fix issue where flow targets built incorrect binary

* packaging: create grafana-agent subdirectory

Create a packaging/grafana-agent subdirectory where Grafana
Agent-specific packaging code is held. This will make it easier to
create packaging for grafana-agent-flow.

* packaging: add grafana-agent-flow package

This commit adds a grafana-agent-flow system package for DEB and RPM
systems.

Closes #3303.
