+++ .drone/drone.yml
@@ -1246,6 +1246,7 @@ steps:
 - commands:
   - DOCKER_OPTS="" make dist/grafana-agent-linux-amd64
   - DOCKER_OPTS="" make dist/grafana-agentctl-linux-amd64
+  - DOCKER_OPTS="" make dist.temp/grafana-agent-flow-linux-amd64
   - DOCKER_OPTS="" make test-packages
   image: grafana/agent-build-image:0.23.0
   name: Test Linux system packages
@@ -1301,6 +1302,6 @@ kind: secret
 name: gpg_passphrase
 ---
 kind: signature
-hmac: 1193d183fa9ceda6303a8f70a874dc3f0b76f9f59911ac921054ccb4385c801f
+hmac: eedbec33572c4b8af8ea0809651674b90813decd930c0ee994e35ecb45d06c40
 
 ...
