+++ CHANGELOG.md
@@ -40,9 +40,12 @@ Main (unreleased)
   - `prometheus.exporter.mysql` collects metrics from a MySQL database. (@spartan0x117)
   - `prometheus.exporter.postgres` collects metrics from a PostgreSQL database. (@spartan0x117)
   - `prometheus.exporter.snmp` collects metrics from SNMP exporter (@marctc).
-  - `otelcol.auth.sigv4` performs AWS Signature Version 4 (SigV4) authentication 
+  - `otelcol.auth.sigv4` performs AWS Signature Version 4 (SigV4) authentication
     for making requests to AWS services via `otelcol` components that support
     authentication extensions. (@ptodev)
+
+- Add support for Flow-specific DEB and RPM system packages. This allows users
+  to install Grafana Agent Flow alongside Grafana Agent. (@rfratto, @robigan)
 
 ### Enhancements
