+++ packaging/grafana-agent-flow/deb/control/prerm
@@ -1,0 +1,12 @@
+#!/bin/sh
+
+set -e
+
+# shellcheck disable=SC1091
+[ -f /etc/default/grafana-agent-flow ] && . /etc/default/grafana-agent-flow
+
+if [ "$1" = "remove" ]; then
+    if command -v systemctl 2>/dev/null; then
+        systemctl stop grafana-agent-flow.service > /dev/null 2>&1 || :
+    fi
+fi
