+++ packaging/grafana-agent-flow/environment-file
@@ -1,0 +1,16 @@
+## Path:
+## Description: Grafana Agent Flow settings
+## Type:        string
+## Default:     ""
+## ServiceRestart: grafana-agent-flow
+#
+# Command line options for grafana-agent
+#
+# The configuration file holding the agent config.
+CONFIG_FILE="/etc/grafana-agent-flow.river"
+
+# User-defined arguments to pass to the run command.
+CUSTOM_ARGS=""
+
+# Restart on system upgrade. Defaults to true.
+RESTART_ON_UPGRADE=true
