+++ packaging/grafana-agent-flow/rpm/control/prerm
@@ -1,0 +1,20 @@
+#!/bin/sh
+
+set -e
+
+# shellcheck disable=SC1091
+[ -f /etc/sysconfig/grafana-agent-flow ] && . /etc/sysconfig/grafana-agent-flow
+
+# final uninstallation $1=0
+# If other copies of this RPM are installed, then $1>0
+
+if [ "$1" -eq 0 ] ; then
+    if [ -x /bin/systemctl ] ; then
+        /bin/systemctl stop grafana-agent-flow.service > /dev/null 2>&1 || :
+    elif [ -x /etc/init.d/grafana-agent-flow ] ; then
+        /etc/init.d/grafana-agent-flow stop
+    elif [ -x /etc/rc.d/init.d/grafana-agent-flow ] ; then
+        /etc/rc.d/init.d/grafana-agent-flow stop
+    fi
+fi
+exit 0
