+++ packaging/grafana-agent-flow/rpm/grafana-agent-flow.service
@@ -1,0 +1,19 @@
+[Unit]
+Description=Vendor-neutral programmable observability pipelines.
+Documentation=https://grafana.com/docs/agent/latest/flow/
+Wants=network-online.target
+After=network-online.target
+
+[Service]
+Restart=always
+User=grafana-agent
+Environment=HOSTNAME=%H
+EnvironmentFile=/etc/default/grafana-agent-flow
+WorkingDirectory=/var/lib/grafana-agent-flow
+ExecStart=/usr/bin/grafana-agent-flow run $CUSTOM_ARGS --storage.path=/var/lib/grafana-agent-flow $CONFIG_FILE
+ExecReload=/usr/bin/kill -HUP $MAINPID
+TimeoutStopSec=20s
+SendSIGKILL=no
+
+[Install]
+WantedBy=multi-user.target
