+++ packaging/grafana-agent/windows/.gitignore
@@ -1,0 +1,2 @@
+*.exe
+LICENSE
