+++ packaging/grafana-agent/windows/Dockerfile
