+++ packaging/grafana-agent/windows/install_script.nsis
