+++ packaging/grafana-agent/windows/logo.ico
